Fix dtype in test

+++ tests/geometry/geodesic_test.py
@@ -131,7 +131,7 @@ class TestGeodesic:
   def test_approximates_ground_truth(self, t: Optional[float], order: int):
     tol = 1e-2
     G = nx.linalg.adjacency_matrix(balanced_tree(r=2, h=5))
-    G = jnp.asarray(G.toarray())
+    G = jnp.asarray(G.toarray().astype(float))
     eye = jnp.eye(G.shape[0])
     eps = jnp.finfo(eye.dtype).tiny
